Attempt at fixing PHP8 tests in CI
